% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.getMSEA_Metabolon.r,
%   R/data.getORA_Metabolon.r
\name{data.getMSEA_Metabolon}
\alias{data.getMSEA_Metabolon}
\title{Metabolite set enrichment analysis (MSEA) using pathway knowledge curated by Metabolon}
\usage{
data.getMSEA_Metabolon(met.profile, threshold = 3, type = "zscore",
  gene.profile = NULL)

data.getMSEA_Metabolon(met.profile, threshold = 3, type = "zscore",
  gene.profile = NULL)
}
\arguments{
\item{met.profile}{- A character vector of a patient's metabolomic profile, including KEGG IDs
and the associated z-score or p-value describing the level of the metabolite compared to controls.}

\item{threshold}{- A cutoff to select metabolites with a zscore > threshold or < -1*threshold.}

\item{type}{- Either "p-value" or "z-score".}

\item{gene.profile}{- Default set to NULL, meaning the default enrichment analysis only considers metabolites. However, if you have gene data, too, set
this parameter to a character vector of the gene names with found variants in the patient's record. Gene IDs must be converted to Entrez Identifiers.}

\item{p.corr}{- A vector of correlations representing the correlation of the metabolite level with a disease phenotype.
The names of the vector are the metabolite names.}

\item{met.profile}{- A character vector of a patient's metabolomic profile, including KEGG IDs
and the associated z-score or p-value describing the level of the metabolite compared to controls.}
}
\description{
A function that returns the pathway enrichment score for all perturbed metabolites in a patient's full metabolomic profile.

A function that returns the pathway enrichment score for all perturbed metabolites in a patient's full metabolomic profile.
}
\examples{
data(Miller2015_Heparin)
diagnoses = gsub("[[:digit:]]", "", colnames(Miller2015_Heparin))
diag.ind = diagnoses
diag.ind[which(diag.ind!="Argininemia")] = 0
diag.ind[which(diag.ind=="Argininemia")] = 1
diag.ind = as.numeric(diag.ind)
profile.ind = which(diagnoses=="Argininemia")[1]
met.profile = Miller2015_Heparin[,profile.ind]
names(met.profile) = rownames(Miller2015_Heparin)

population = names(met.profile)
paths.hsa = list.dirs(path="../inst/extdata", full.names = FALSE)
paths.hsa = paths.hsa[-which(paths.hsa \%in\% c("", "RData", "allPathways"))]
sink("MetaboliteSetDatabases/Miller2015.gmt")
for (p in 1:length(paths.hsa)) {
  load(sprintf("../inst/extdata/RData/\%s.RData", paths.hsa[p]))
  pathway.compounds = V(ig)$label[which(V(ig)$shape=="circle")]
  pathCompIDs = unique(tolower(pathway.compounds[which(pathway.compounds \%in\% population)]))
  print(sprintf("\%s         \%s", paths.hsa[p], paste(pathCompIDs, collapse="    ")), quote=FALSE)
}
sink()
print("test")
abs_filename_dataset = "Datasets/Miller2015.gct"
abs_filename_classes = "Datasets/Miller2015_arg.cls"
pathway.data = data.getMSEA_Metabolon(met.profile, diag.ind, Miller2015_Heparin)

pathway.data = data.getMSEA_Metabolon(met.profile, threhold=3, "z-score", NULL)
}
