% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.HMDBtoKEGG.r
\name{data.HMDBtoKEGG}
\alias{data.HMDBtoKEGG}
\title{Convert HMDB IDs to KEGG compound IDs.}
\usage{
data.HMDBtoKEGG(hmdb.ids)
}
\arguments{
\item{hmdb.ids}{- A character vector of HMDB IDs.}
}
\description{
A function that converts HMDB IDs to KEGG compounds.
}
\examples{
data.kegg = data.HMDBtoKEGG(data)
}
