% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.LabeltoKEGG.r
\name{data.LabeltoKEGG}
\alias{data.LabeltoKEGG}
\title{Convert compound names to KEGG compound IDs.}
\usage{
data.LabeltoKEGG(compound.names)
}
\arguments{
\item{compound.names}{- A character vector of metabolite (compound) names.}
}
\description{
A function that converts HMDB IDs to KEGG compounds.
}
\examples{
data.kegg = data.HMDBtoKEGG(data)
}
