% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPatientReport.r
\name{getPatientReport}
\alias{getPatientReport}
\title{Get Patient Report}
\usage{
getPatientReport(input, all_data)
}
\arguments{
\item{all_data}{- The entire data matrix loaded based on the diagnosis selected in the dropdown menu input$diagClass.}

\item{PatientID}{- The patient identifier string associated with the patient's profile.}
}
\value{
patientReport - a data table with the metabolites and z-scores associated with the selected patient ID.
}
\description{
Get Patient Report
}
