% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPathwayIgraph.r
\name{getPathwayIgraph}
\alias{getPathwayIgraph}
\title{getPathwayIgraph}
\usage{
getPathwayIgraph(input, Pathway.Name)
}
\arguments{
\item{Pathway.Name}{- The name of the pathway map for which you want the topological information.}
}
\description{
getPathwayIgraph
}
