% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPatientReport.r
\name{getPatientReport}
\alias{getPatientReport}
\title{Get Patient Report}
\usage{
getPatientReport(input, all_data)
}
\arguments{
\item{all_data}{- The entire data matrix loaded based on the diagnosis selected in the dropdown menu input$diagClass.}

\item{PatientID}{- The patient identifier string associated with the patient's profile.}
}
\value{
patientReport - a data table with the metabolites and z-scores associated with the selected patient ID.
}
\description{
Get Patient Report
}
\examples{
data(Miller2015_Heparin)
# Input is supplied by R shiny app, but you can hard code parameters as a list object, too, to test functionality.
input = list()
input$ptIDs = colnames(Miller2015_Heparin)[4]
input$diagClass = "paa"
rpt = getPatientReport(input, Miller2015_Heparin)
head(rpt$patientReport)
}
